+++ .github/workflows/deploy.yml
@@ -7,7 +7,7 @@ name: Deploy Next.js site to Pages
 on:
   # Runs on pushes targeting the default branch
   push:
-    branches: ["main"]
+    branches: ["*"] # update to main for production
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
