+++ frontend/package.json
@@ -36,6 +36,7 @@
     "react-markdown": "^9.0.1",
     "react-resizable-panels": "^2.1.7",
     "redis": "^4.7.0",
+    "serve": "^14.2.4",
     "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "tailwind-merge": "^2.5.5",
